+++ packages/test-utils/src/helpers/react-testing-library/comboboxHelpers.ts
@@ -24,7 +24,17 @@
 
  */
 
-export { ReactWrapper } from 'enzyme'
-export * from './create_with_theme'
-export * from './snapshot'
-export * from './helpers/react-testing-library'
+import { fireEvent, getNodeText, screen } from '@testing-library/react'
+
+export const openCombobox = (placeholderText: string) =>
+  fireEvent.mouseDown(screen.getByPlaceholderText(placeholderText))
+
+export const closeCombobox = () => fireEvent.click(document)
+
+export const getComboboxOptions = () => screen.getAllByRole('option')
+
+export const getComboboxOptionText = (el: HTMLElement) =>
+  getNodeText(el.children[1] as HTMLElement)
+
+export const getAllComboboxOptionText = () =>
+  getComboboxOptions().map(getComboboxOptionText)