+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ActionListItem` accepts `actionsButtonLabel` prop to help with testing
 - `ComboboxList` and `ComboboxOption` now both support a custom `indicator`
   - `Select` and `SelectMulti` also support `indicator` at both the component and option level
+- Test helper functions for `Combobox` component using react-testing-library
 
 ### Changed